requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
alembic>=1.7.1
python-multipart>=0.0.5
requests>=2.26.0
beautifulsoup4
psycopg2-binary
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=0.19.0